resource-server/src/main/java/com/example/web/controller/FooController.java
package com.example.web.controller;

import com.example.web.dto.Foo;
import com.example.web.service.IFooService;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.ResponseBody;

@Controller
public class FooController {

    private final IFooService _fooService;

    public FooController(final IFooService fooService) {
        _fooService = fooService;
    }

    @PreAuthorize("#oauth2.hasScope('read')")
    @RequestMapping(method = RequestMethod.GET, value = "/foos/{id}")
    @ResponseBody
    public Foo findById(@PathVariable long id) {
        return _fooService.findById(id);
    }
}
